backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
celery==5.4.0
redis==5.0.8
groq==0.11.0
httpx==0.27.2
python-dotenv==1.0.1
slowapi==0.1.9
loguru==0.7.2
motor==3.6.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
rank-bm25==0.2.2
langchain-experimental==0.4.1
ultralytics==8.4.5
supabase==2.10.0
pdfplumber==0.11.4
python-docx==1.1.2
langchain-text-splitters==0.3.2
Pillow==10.1.0
python-magic==0.4.27
easyocr==1.7.1
numpy==1.26.4
torch==2.4.1
torchvision==0.19.1
certifi==2025.11.12
